Fix exponential blowup with blocks with `If`
Previously, the functions `TealBlock.addIncoming` and
`TealBlock.validate` took exponential time with blocks with `If`.

The issue is that `If` blocks have two children (`left` and `right`),
and both of them have the same `next` child block.

To see the issue, run `examples/signature/test_many_ifs.py`
before and after this PR.