Modify ABI DynamicBytes and StaticBytes to support router/subroutine
Router/subroutine requires that an associated TypeSpec
to act as a factory for those types.

This commit is meant to add this associated TypeSpec.

Note: I am not very familiar with the code and may
have made mistakes. However, unit tests work and the code
works as expected on a project I'm working on.